kas: Add kas-sulka-configuration
Add a new kas-sulka-configuration file that allows defining the
local.conf configuration items. This can be used to set the service user
name, password, or to enable graphics support. This file can also be
used for configuration items not specific to Sulka.

By default, set the service user password to empty. This is the default
functionality in the distro configuration, but allows easier
modifications from the developers when it is already in the
configuration file.