extra_fragments: Add extra config fragments
Add extra configuration yml fragments that can be used to edit the
local.conf. This can be used to configure the build for various
purposes on top of the hardened Sulka image.

Add two fragments: audit fragment that install lynis and openscap, and a
development fragment that installs SSH server.