kas: Add ordering to local.conf headers
Add ordering to local.conf headers in the configuration files. This
allows ensuring that the overrides, etc. can be easily defined in the
correct order.

For the existing yml files, ensure that the kas-sulka-configuration is
the base, audit configuration can modify this base if required, and
development configuration can modify everything.