scripts: Add script to generate SSH keys
Add script that can be used to generate SSH keys. The public key is
generated in a naming format that it can be used in build to add it to
the image. The script allows choosing the desired location for the keys.
This allows easier key generation for the images for development
purposes.